Remove duplicate function between the blockchains
