+++ blockchain/src/blockchain/slots.rs
@@ -1,11 +1,10 @@
-use nimiq_blockchain_interface::BlockchainError;
-use nimiq_collections::BitSet;
+use nimiq_blockchain_interface::{AbstractBlockchain, BlockchainError};
 use nimiq_database::TransactionProxy;
 use nimiq_primitives::{
     policy::Policy,
     slots_allocation::{Slot, Validators},
 };
-use nimiq_vrf::{Rng, VrfEntropy, VrfSeed, VrfUseCase};
+use nimiq_vrf::{VrfEntropy, VrfSeed};
 
 use crate::Blockchain;
 
@@ -72,7 +71,11 @@ impl Blockchain {
             .next_batch_initial_punished_set;
 
         // Compute the slot number of the next proposer.
-        let slot_number = Self::compute_slot_number(offset, vrf_entropy, disabled_slots);
+        let slot_number = <Blockchain as AbstractBlockchain>::compute_slot_number(
+            offset,
+            vrf_entropy,
+            disabled_slots,
+        );
 
         // Fetch the validators that are active in given block's epoch.
         let epoch_number = Policy::epoch_at(block_number);
@@ -89,32 +92,5 @@ impl Blockchain {
             band: slot_band,
             validator: validator.clone(),
         })
-    }
-
-    fn compute_slot_number(offset: u32, vrf_entropy: VrfEntropy, disabled_slots: BitSet) -> u16 {
-        // RNG for slot selection
-        let mut rng = vrf_entropy.rng(VrfUseCase::ViewSlotSelection);
-
-        // Create a list of viable slots.
-        let mut slots: Vec<u16> = if disabled_slots.len() == Policy::SLOTS as usize {
-            // If all slots are disabled, we will accept any slot, since we want the
-            // chain to progress.
-            (0..Policy::SLOTS).collect()
-        } else {
-            // Otherwise, we will only accept slots that are not disabled.
-            (0..Policy::SLOTS)
-                .filter(|slot| !disabled_slots.contains(*slot as usize))
-                .collect()
-        };
-
-        // Shuffle the slots vector using the Fisher–Yates shuffle.
-        for i in (1..slots.len()).rev() {
-            let r = rng.next_u64_below((i + 1) as u64) as usize;
-            slots.swap(r, i);
-        }
-
-        // Now simply take the offset modulo the number of viable slots and that will give us
-        // the chosen slot.
-        slots[offset as usize % slots.len()]
     }
 }
